refactor: login register code refactoring
