+++ src/controllers/users.js
@@ -3,48 +3,6 @@ import { AccountsService } from '../services/accounts.js';
 import { UsersService } from '../services/users.js';
 
 export class UsersController {
-  static async register(req, res, next) {
-    try {
-      const data = req.body;
-
-      const userRegister = await UsersService.register(data);
-
-      delete userRegister.user.password;
-
-      res.json({
-        meta: {
-          statusCode: 200,
-          message: 'register successfully',
-        },
-        data: userRegister,
-      });
-    } catch (error) {
-      next(error);
-    }
-  }
-
-  static async login(req, res, next) {
-    try {
-      const data = req.body;
-
-      const { user, token } = await UsersService.login(data);
-
-      delete user.password;
-
-      res.json({
-        meta: {
-          statusCode: 200,
-          message: 'login successfully',
-        },
-        data: {
-          _token: token,
-        },
-      });
-    } catch (error) {
-      next(error);
-    }
-  }
-
   static async getUserLoggedIn(req, res, next) {
     try {
       const userLoggedIn = req.user;
